% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_bandpass_filter.R
\name{apply_bandpass_filter}
\alias{apply_bandpass_filter}
\title{Apply bandpass filter to EEG signal}
\usage{
apply_bandpass_filter(
  signal,
  low_freq = 0.5,
  high_freq = 30,
  sampling_rate = 256
)
}
\arguments{
\item{signal}{The long EEG matrix}

\item{low_freq}{the lower cutoff frequency}

\item{high_freq}{the upper cutoff frequency}

\item{sampling_rate}{the sampling rate of the EEG data}
}
\value{
EEG data after applying the bandpass filter. It has the same dimension as \code{signal}.
}
\description{
Apply bandpass filter to EEG signal
}
