% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eegmat.R
\name{create_eegmat}
\alias{create_eegmat}
\title{Helper function to convert the long EEG matrix into a tabular form}
\usage{
create_eegmat(signal, flash_loc, nT = 205, downrate = 1)
}
\arguments{
\item{signal}{The long EEG signal matrix.}

\item{flash_loc}{Locations to extract EEG segments.}

\item{nT}{How many time points to extract.}

\item{downrate}{Downsampling rate.}
}
\value{
A \code{length(flash_loc)} by \code{nT * ncol(signal)} EEG matrix.
}
\description{
Helper function to convert the long EEG matrix into a tabular form
}
