% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_accuracy.R
\name{evaluate_accuracy}
\alias{evaluate_accuracy}
\title{Helper function to evaluate prediction accuracy}
\usage{
evaluate_accuracy(flashes, score, cutoff = 0.5)
}
\arguments{
\item{flashes}{a list containing flash-level information. Should have
TrialNR, StimulusType, and StimulusCode.}

\item{score}{prediction scores for each flash.}

\item{cutoff}{for flash-level accuracy only.
A flash with score > cutoff will be predicted as target,}
}
\value{
The flash-level accuracy and character-level accuracy
}
\description{
Currently, this function only works for K protocol data
}
