% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dat.R
\name{load_dat}
\alias{load_dat}
\title{Load .mat files to R}
\usage{
load_dat(file_path)
}
\arguments{
\item{file_path}{The file path of the .mat file}
}
\value{
A nested list that represents the .mat file
}
\description{
Load .mat files to R
}
