% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify}
\alias{simplify}
\title{Simplify the raw data to focus on less imformation}
\usage{
simplify(dat)
}
\arguments{
\item{dat}{The nested list of one run}
}
\value{
A simplier nested list that contains important information
\itemize{
\item \code{signal}: The long EEG signal matrix.
\item \code{states}: State variables.
\item \code{additional_info}: Other useful variables such as
sampling frequency and channel names.
}
}
\description{
Simplify the raw data to focus on less imformation
}
