% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_eegmat.R
\name{create_eegmat}
\alias{create_eegmat}
\title{Helper function to convert the long EEG matrix into a tabular form}
\usage{
create_eegmat(signal, flash_loc, nT = 205)
}
\arguments{
\item{signal}{The long EEG signal matrix of dim # of time points by # channels.
For example, \code{dat[["Data"]][["RawData"]][["signal"]]}.}

\item{flash_loc}{Locations of flashes corresponding to row indexes in \code{signal}.}

\item{nT}{How many time points to extract after each flash.}
}
\value{
A \code{length(flash_loc)} by \code{nT} EEG matrix.
}
\description{
Helper function to convert the long EEG matrix into a tabular form
}
