% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_info.R
\name{extract_info}
\alias{extract_info}
\title{Helper function to extract useful information}
\usage{
extract_info(dat)
}
\arguments{
\item{dat}{Output from \code{load_dat}}
}
\value{
A list
\itemize{
\item{\code{flash_loc}: }{The starting point of each flash. }
\item{\code{flash_type}: }{Whether the flash is a target, same length as \code{flash_loc}.
Target as 1 and non-target as 0.}
\item{\code{code}: }{The flash code, same length as \code{flash_loc},
1-6 means rows and 7-12 means columns.}
\item{\code{nfpc}: }{Number of flash generated per each character. For example,
if one character has four sequences, then \code{nfpc = 48}, because
each sequence has 12 flashes}
}
}
\description{
Helper function to extract useful information
}
